requirements.txt
absl-py==1.3.0
astor==0.8.1
atari-py==0.2.9
attrs==22.2.0
cached-property==1.5.2
cloudpickle==1.6.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
gast==0.5.3
google-pasta==0.2.0
grpcio==1.48.2
gym==0.17.3
h5py==3.1.0
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
natsort==8.2.0
networkx==2.5.1
numpy==1.19.5
opencv-python==4.4.0.46
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyglet==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
stable-baselines==2.10.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tqdm==4.64.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
